+++ strucmotif-search-core/src/main/java/org/rcsb/strucmotif/io/InvertedIndexImpl.java
@@ -4,6 +4,7 @@ import org.rcsb.strucmotif.config.InvertedIndexBackend;
 import org.rcsb.strucmotif.config.StrucmotifConfig;
 import org.rcsb.strucmotif.domain.bucket.Bucket;
 import org.rcsb.strucmotif.domain.motif.ResiduePairIdentifier;
+import org.rcsb.strucmotif.domain.structure.PolymerType;
 import org.rcsb.strucmotif.io.codec.BucketCodec;
 import org.rcsb.strucmotif.domain.bucket.ResiduePairIdentifierBucket;
 import org.rcsb.strucmotif.domain.bucket.InvertedIndexBucket;
@@ -27,12 +28,16 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.Collection;
+import java.util.Collections;
 import java.util.HashMap;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
+import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.function.Function;
+import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import java.util.zip.GZIPInputStream;
@@ -51,6 +56,7 @@ public class InvertedIndexImpl implements InvertedIndex {
     private final boolean gzipped;
     private final String extension;
     private final BucketCodec bucketCodec;
+    private final Pattern tempExtension;
     private boolean paths;
 
     /**
@@ -64,22 +70,105 @@ public class InvertedIndexImpl implements InvertedIndex {
         this.bucketCodec = backend.getBucketCodec();
         this.extension = backend.getExtension() + (gzipped ? ".gz" : "");
         logger.info("Index files will {}be gzipped - extension: {}", gzipped ? "" : "not ", extension);
+        this.tempExtension = Pattern.compile(".*" + extension.replace(".", "\\.") + "\\.[0-9]+$");
         this.paths = false;
     }
 
     @Override
-    public void insert(ResiduePairDescriptor residuePairDescriptor, Bucket bucket) {
+    public void insert(ResiduePairDescriptor residuePairDescriptor, Bucket bucket, int batchId) {
         if (!paths) {
             ensureDirectoriesExist();
             this.paths = true;
         }
 
         try {
-            Path path = getPath(residuePairDescriptor);
-            ResiduePairIdentifierBucket merged = merge(getBucket(residuePairDescriptor), bucket);
+            // write a temporary file (appended by the batchId)
+            Path tmpPath = getPath(residuePairDescriptor, batchId);
+            try (ByteArrayOutputStream outputStream = bucketCodec.encode(bucket)) {
+                write(tmpPath, outputStream);
+            }
+        } catch (IOException e) {
+            throw new UncheckedIOException(e);
+        }
+    }
 
-            try (ByteArrayOutputStream outputStream = bucketCodec.encode(merged)) {
-                write(path, outputStream);
+    @Override
+    public void commit() {
+        logger.info("Committing temporary files to index");
+        try {
+            Map<Path, Set<Path>> toMerge = new ConcurrentHashMap<>();
+            try (Stream<Path> paths = temporaryFiles()) {
+                AtomicInteger counter = new AtomicInteger();
+                paths.peek(p -> progress(counter, 10000, "{} files scanned")).forEach(p -> {
+                    String filename = p.getFileName().toString();
+                    Path persistentPath = p.resolveSibling(filename.substring(0, filename.lastIndexOf(".")));
+                    Set<Path> bin = toMerge.computeIfAbsent(persistentPath, e -> Collections.synchronizedSet(new HashSet<>()));
+                    bin.add(persistentPath.resolveSibling(p));
+                });
+            }
+
+            logger.info("Merging {} bins", toMerge.size());
+
+            toMerge.entrySet()
+                    .parallelStream()
+                    .forEach(entry -> {
+                        Path destination = entry.getKey();
+                        Set<Path> sources = entry.getValue();
+
+                        try {
+                            Map<Integer, Collection<ResiduePairIdentifier>> merged = new HashMap<>();
+                            // populate with persistent data
+                            if (Files.exists(destination)) {
+                                try (InputStream inputStream = getInputStream(destination)) {
+                                    addAll(merged, bucketCodec.decode(inputStream),  null);
+                                }
+                            } else {
+                                // there's only a single file and the destination exists not yet (aka no merging to be done)
+                                if (sources.size() == 1) {
+                                    Path source = sources.stream().findFirst().orElseThrow();
+                                    Files.move(source, destination);
+                                    logger.debug("Moved {} to {}", source, destination);
+                                    return;
+                                }
+                            }
+
+                            // merge all new, temporary data
+                            for (Path p : sources) {
+                                try (InputStream inputStream = getInputStream(p)) {
+                                    addAll(merged, bucketCodec.decode(inputStream), null);
+                                }
+                            }
+                            ByteArrayOutputStream outputStream = bucketCodec.encode(new ResiduePairIdentifierBucket(merged));
+                            write(destination, outputStream);
+                            outputStream.close();
+
+                            for (Path source : sources) {
+                                Files.delete(source);
+                            }
+                            logger.debug("Merged {} into {}", sources, destination);
+                        } catch (IOException e) {
+                            throw new UncheckedIOException(e);
+                        }
+                    });
+        } catch (IOException e) {
+            throw new UncheckedIOException(e);
+        }
+    }
+
+    @Override
+    public void clearTemporaryFiles() {
+        try {
+            logger.info("Collecting all temporary index files at {}", basePath);
+            AtomicInteger counter = new AtomicInteger();
+            try (Stream<Path> paths = temporaryFiles()) {
+                paths.peek(p -> progress(counter, 10000, "{} files deleted"))
+                        .forEach(p -> {
+                            try {
+                                Files.delete(p);
+                            } catch (IOException e) {
+                                throw new UncheckedIOException(e);
+                            }
+                        });
             }
         } catch (IOException e) {
             throw new UncheckedIOException(e);
@@ -112,7 +201,10 @@ public class InvertedIndexImpl implements InvertedIndex {
      * @throws IOException reading failed
      */
     protected InputStream getInputStream(ResiduePairDescriptor residuePairDescriptor) throws IOException {
-        Path path = getPath(residuePairDescriptor);
+        return getInputStream(getPath(residuePairDescriptor));
+    }
+
+    private InputStream getInputStream(Path path) throws IOException {
         InputStream inputStream = Files.newInputStream(path);
         return gzipped ? new GZIPInputStream(inputStream, BUFFER_SIZE) : new BufferedInputStream(inputStream, BUFFER_SIZE);
     }
@@ -121,6 +213,11 @@ public class InvertedIndexImpl implements InvertedIndex {
         String bin = residuePairDescriptor.toString();
         String uberbin = bin.substring(0, 2);
         return basePath.resolve(uberbin).resolve(bin + extension);
+    }
+
+    private Path getPath(ResiduePairDescriptor residuePairDescriptor, int batchId) {
+        Path actual = getPath(residuePairDescriptor);
+        return actual.resolveSibling(actual.getFileName().toString() + "." + batchId);
     }
 
     private InvertedIndexBucket getBucket(ResiduePairDescriptor residuePairDescriptor) {
@@ -191,6 +288,8 @@ public class InvertedIndexImpl implements InvertedIndex {
     private void ensureDirectoriesExist() {
         try {
             List<String> oneLetterCodes = Stream.of(ResidueType.values())
+                    // ignore unknown component
+                    .filter(residueType -> residueType.getPolymerType() != PolymerType.UNKNOWN_POLYMER)
                     .map(ResidueType::getInternalCode)
                     // has to be sorted to honor the implicit contract on when identifiers are flipped
                     .sorted()
@@ -243,37 +342,35 @@ public class InvertedIndexImpl implements InvertedIndex {
     }
 
     private Stream<Path> indexFiles() throws IOException {
+        return files().filter(p -> p.getFileName().toString().endsWith(extension));
+    }
+
+    private Stream<Path> temporaryFiles() throws IOException {
+        return files().filter(p -> tempExtension.matcher(p.getFileName().toString()).matches());
+    }
+
+    private Stream<Path> allFiles() throws IOException {
+        return files().filter(p -> p.getFileName().toString().contains(extension));
+    }
+
+    private Stream<Path> files() throws IOException {
         if (!Files.exists(basePath)) {
             return Stream.empty();
         }
 
         return Files.walk(basePath, FileVisitOption.FOLLOW_LINKS)
                 .parallel()
-                // ignore directories
-                .filter(p -> !Files.isDirectory(p) && p.getFileName().toString().contains(extension));
+                .filter(Files::isRegularFile);
     }
 
-    /**
-     * Merge two buckets. Doesn't allow duplicates.
-     * @param bucket1 source1
-     * @param bucket2 source2
-     * @return a new bucket that contains the combined results of both
-     */
-    private ResiduePairIdentifierBucket merge(Bucket bucket1, Bucket bucket2) {
-        Map<Integer, Collection<ResiduePairIdentifier>> map = new HashMap<>();
-        addAll(map, bucket1, false, null);
-        addAll(map, bucket2, true, null);
-        return new ResiduePairIdentifierBucket(map);
-    }
-
-    private void addAll(Map<Integer, Collection<ResiduePairIdentifier>> map, Bucket bucket, boolean noDuplicates, Collection<Integer> ignore) {
+    private void addAll(Map<Integer, Collection<ResiduePairIdentifier>> map, Bucket bucket, Collection<Integer> ignore) {
         while (bucket.hasNextStructure()) {
             bucket.moveStructure();
             int key = bucket.getStructureIndex();
             if (ignore != null && ignore.contains(key)) {
                 continue;
             }
-            if (noDuplicates && map.containsKey(key)) {
+            if (map.containsKey(key)) {
                 throw new IllegalStateException("Duplicate key: " + key);
             }
             Collection<ResiduePairIdentifier> identifiers = map.computeIfAbsent(key, e -> new ArrayList<>());
@@ -295,7 +392,7 @@ public class InvertedIndexImpl implements InvertedIndex {
      */
     private ResiduePairIdentifierBucket removeByKey(InvertedIndexBucket bucket, Collection<Integer> removals) {
         Map<Integer, Collection<ResiduePairIdentifier>> map = new HashMap<>();
-        addAll(map, bucket, true, removals);
+        addAll(map, bucket, removals);
         return new ResiduePairIdentifierBucket(map);
     }
 }