+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Unreleased
 ### Added
 - support for modified residues
 - support for pyrrolysine, selenocysteine, inosinic acid & 2'-deoxyuridine-5'-monophosphate
+- write index updates as temporary files and merge only as final operation
 
 ### Bug fix
 - add 2'-Deoxythymidine-5'-monophosphate as `DT`